--- data/duke.txt
@@ -1,1 +1,0 @@
-E | 0 | return book | 2/12/2019 1800
